pinot-dashboard/requirements.txt
flake8==2.4.0
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
kazoo==2.0
MarkupSafe==0.23
mccabe==0.3
pep8==1.5.7
pyflakes==0.8.1
PyYAML==3.11
requests==2.5.3
Werkzeug==0.10.1
